git-svn-id: http://www.ifountain.org/repos/os/RapidModules@3029 a08bcbf9-5f6c-4d33-af0a-e494bc2cacca

+++ RapidCMDB/grails-app/domain/datasource/DatabaseDatasource.groovy
@@ -9,7 +9,7 @@ class DatabaseDatasource extends BaseDatasource{
 
     
     DatabaseConnection connection ;
-    int reconnectInterval = 0;
+    Long reconnectInterval = 0;
     
 
     static relations = [
